rtc: alarm: update power off alarm setting logic
The previous power_on_alarm value should be checked to avoid
cancelling the alarm by mistake.And currently we setup the
RTC for poweroff alarm when the system performs shutdown. But
if the phone is shut down unexpectedly before that, the
poweroff alarm can't work. So we can do that immediately.

Change-Id: I1749abfae546850745659cbea43a16b1abf01f35
Signed-off-by: Xiaocheng Li <lix@codeaurora.org>